pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prismllm"
version = "0.1.0"
description = "Progressive LLM Data Platform"
authors = [{name = "SparkSnail", email = "yangshinai@gmail.com"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
keywords = ["llm", "ai", "machine-learning", "vllm", "rag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[project.scripts]
prismllm = "prismllm.cli:main"

[project.urls]
Homepage = "https://github.com/SparkSnail/PrismLLM"
Repository = "https://github.com/SparkSnail/PrismLLM"
